Add single post template and style
A new template for displaying single posts was created and integrated into the site's main theme. This includes the addition of navigation for previous and next posts, and conditional loading of the comment template. On the Styling aspect, support for this single post view was added to the project's SCSS files.